+++ docs/CHANGELOG.md
@@ -1,13 +1,28 @@
 Changelog
 =========
+
+0.9.20
+-------------------------
+- Enh #15: Implement user profile polls
+- Fix: Poll create form rendered for users without create poll permission
+- Fix #59: Only public/private poll stream filter not working 
+
+0.9.19 - October 16, 2019
+-------------------------
+- Fix: 1.4 compatibility removed deprecated content usage
+- Enh: 1.4 csp nonce support
+- Enh: Raised minimum HumHub version to 1.3 
+
+
 0.9.18 - June 17, 2019
+-----------------------
 - Enh: Updated translations
 - Enh: Updated docs
 
 
 0.9.17 - November 07, 2018
------------------------
-- Fix: Empty answer submittion not intercepted
+--------------------------
+- Fix: Empty answer submission not intercepted
 
 
 0.9.16 - July 25, 2018